devtools/client/webide/test/test_zoom.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        (async function() {
          let win = await openWebIDE();
          let viewer = win.QueryInterface(Ci.nsIInterfaceRequestor)
                            .getInterface(Ci.nsIWebNavigation)
                            .QueryInterface(Ci.nsIDocShell)
                            .contentViewer;

          win.Cmds.zoomOut();
          win.Cmds.zoomOut();
          win.Cmds.zoomOut();
          win.Cmds.zoomOut();
          win.Cmds.zoomOut();
          win.Cmds.zoomOut();
          win.Cmds.zoomOut();

          let roundZoom = Math.round(10 * viewer.fullZoom) / 10;
          is(roundZoom, 0.6, "Reach min zoom");

          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();
          win.Cmds.zoomIn();

          roundZoom = Math.round(10 * viewer.fullZoom) / 10;
          is(roundZoom, 1.4, "Reach max zoom");

          await closeWebIDE(win);

          win = await openWebIDE();
          viewer = win.QueryInterface(Ci.nsIInterfaceRequestor)
                        .getInterface(Ci.nsIWebNavigation)
                        .QueryInterface(Ci.nsIDocShell)
                        .contentViewer;

          roundZoom = Math.round(10 * viewer.fullZoom) / 10;
          is(roundZoom, 1.4, "Zoom restored");

          win.Cmds.resetZoom();

          is(viewer.fullZoom, 1, "Zoom reset");

          await closeWebIDE(win);

          SimpleTest.finish();
        })();
      };
    </script>
  </body>
</html>
